# doc/API接口/openapi.yaml
openapi: 3.0.3
info:
  title: 消息通知平台 API
  description: |
    消息通知平台提供统一的消息发送服务，支持多种渠道（邮件、短信、微信等）。
    
    ## 特性
    - 📧 邮件发送（支持HTML、附件、模板）
    - 📱 短信发送（第二期）
    - 💬 微信消息（第二期）
    - 📝 消息模板管理
    - 🔐 JWT Token认证
    - 📊 监控和健康检查
    
    ## 认证方式
    1. 使用API Key和API Secret获取JWT Token
    2. 后续请求携带Token：`Authorization: Bearer {token}`
    
    ## 限流策略
    - 每个API Key默认限制：1000次/小时
    - 超过限制返回429状态码
    
    ## 技术支持
    - 邮箱：support@example.com
    - 文档：https://docs.example.com
    
  version: 1.0.0
  contact:
    name: 技术支持团队
    email: support@example.com
    url: https://docs.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: 本地开发环境
  - url: https://api-dev.example.com
    description: 开发环境
  - url: https://api-staging.example.com
    description: 测试环境
  - url: https://api.example.com
    description: 生产环境
...
paths:
...
  /api/v1/messages/email/send:
    post:
      tags:
        - 邮件
      summary: 发送邮件
      description: |
        发送邮件，支持两种模式：
        1. **直接发送**：提供subject和content
        2. **模板发送**：提供template_id和template_vars
        
        两种模式二选一，不能同时提供。
      operationId: sendEmail
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailSendRequest'
            examples:
              direct:
                summary: 直接发送
                value:
                  to: "user@example.com"
                  subject: "测试邮件"
                  content: "<h1>测试内容</h1>"
                  content_type: "html"
              template:
                summary: 模板发送
                value:
                  to: "user@example.com"
                  template_id: 1
                  template_vars:
                    username: "张三"
                    code: "123456"
                    expire_minutes: "5"
      responses:
        '200':
          description: 邮件发送成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageSendResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: 无可用邮箱账户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/messages/{message_id}:
    get:
      tags:
        - 消息
      summary: 查询消息状态
      description: 根据message_id查询消息发送状态
      operationId: getMessage
      security:
        - bearerAuth: []
      parameters:
        - name: message_id
          in: path
          required: true
          description: 消息ID
          schema:
            type: string
            example: "msg_20251024_abc123"
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/messages:
    get:
      tags:
        - 消息
      summary: 查询消息列表
      description: 分页查询消息列表，支持多种筛选条件
      operationId: listMessages
      security:
        - bearerAuth: []
      parameters:
        - name: recipient
          in: query
          description: 接收人
          schema:
            type: string
        - name: status
          in: query
          description: 消息状态
          schema:
            type: string
            enum: [pending, sending, success, failed]
        - name: start_date
          in: query
          description: 开始日期
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: 结束日期
          schema:
            type: string
            format: date
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
  /api/v1/messages/batch/query:
    post:
      tags:
        - 消息
      summary: 批量查询消息
      description: 根据message_id列表批量查询消息状态
      operationId: batchGetMessages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchQueryRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchQueryResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token认证
  schemas:
...
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          description: 状态码，0表示成功
          example: 0
        message:
          type: string
          description: 响应消息
          example: "success"
        request_id:
          type: string
          description: 请求追踪ID
          example: "req_xyz789"
        timestamp:
          type: string
          format: date-time
          description: 响应时间
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              nullable: true
    
    # ============ 健康检查 ============
...
    EmailSendRequest:
      type: object
      required:
        - to
      properties:
        to:
          type: string
          format: email
          description: 收件人邮箱
          example: "user@example.com"
        cc:
          type: array
          items:
            type: string
            format: email
          description: 抄送列表
        bcc:
          type: array
          items:
            type: string
            format: email
          description: 密送列表
        subject:
          type: string
          description: 邮件主题（直接发送时必填）
          example: "测试邮件"
        content:
          type: string
          description: 邮件内容（直接发送时必填）
          example: "<h1>测试内容</h1>"
        content_type:
          type: string
          enum: [html, text]
          default: html
          description: 内容类型
        template_id:
          type: integer
          description: 模板ID（模板发送时必填）
          example: 1
        template_vars:
          type: object
          description: 模板变量（模板发送时必填）
          additionalProperties: true
          example:
            username: "张三"
            code: "123456"
        attachment_ids:
          type: array
          items:
            type: string
          description: 附件ID列表
          example: ["att_xxx", "att_yyy"]
        idempotency_key:
          type: string
          description: 幂等键，防止重复发送
          example: "unique-key-123"
        async:
          type: boolean
          default: true
          description: 是否异步发送
    MessageSendResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                message_id:
                  type: string
                  example: "msg_20251024_abc123"
                status:
                  type: string
                  enum: [pending, sending, success, failed]
                  example: "pending"
    
    # ============ 消息查询 ============
    MessageDetailResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                message_id:
                  type: string
                message_type:
                  type: string
                  enum: [email, sms, wechat]
                recipient:
                  type: string
                subject:
                  type: string
                status:
                  type: string
                  enum: [pending, sending, success, failed]
                retry_count:
                  type: integer
                error_message:
                  type: string
                  nullable: true
                created_at:
                  type: string
                  format: date-time
                sent_at:
                  type: string
                  format: date-time
                  nullable: true
    MessageListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                total:
                  type: integer
                  example: 100
                page:
                  type: integer
                  example: 1
                page_size:
                  type: integer
                  example: 20
                records:
                  type: array
                  items:
                    type: object
                    properties:
                      message_id:
                        type: string
                      recipient:
                        type: string
                      subject:
                        type: string
                      status:
                        type: string
                      created_at:
                        type: string
                        format: date-time
    BatchQueryRequest:
      type: object
      required:
        - message_ids
      properties:
        message_ids:
          type: array
          items:
            type: string
          maxItems: 100
          example: ["msg_20251024_abc123", "msg_20251024_abc124"]
    BatchQueryResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                type: object
                properties:
                  message_id:
                    type: string
                  status:
                    type: string
                  recipient:
                    type: string
                  sent_at:
                    type: string
                    format: date-time
                    nullable: true
    
    # ============ 模板管理 ============
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1001
            message: "请求参数错误"
            data: null
            request_id: "req_xyz789"
    Unauthorized:
      description: 认证失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1002
            message: "认证失败"
            data: null
            request_id: "req_xyz789"
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1004
            message: "资源不存在"
            data: null
            request_id: "req_xyz789"
security:
  - bearerAuth: []